Agentic_Research_Exploration/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.4.0
python-dotenv
PyPDF2
gradio>=5.0
requests